+++ svelte/CHANGELOG.md
@@ -5,7 +5,8 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
--
+-   improved error message when loading view fails
+-   shows "View not found" on the UI too
 
 ## [0.0.4] - 10/03/2023